Work around faulty diffopt in macOS builtin Vim
macOS builtin Vim seems to be patched to remove support for 'internal'
argument in 'diffopt' (I presume because of the LGPL licence of the
xdiff library used to implement it). But the patch is faulty and is not
updating the default value of the 'diffopt' - it still includes
'internal' - although setting it is an error. This causes most unrelated
changes to 'diffopt' to error.

https://www.micahsmith.com/blog/2019/11/fixing-vim-invalid-argument-diffopt-iwhite/